renaming http2 to fetch, polishing code
