+++ pnpm-workspace.yaml
@@ -16,5 +16,6 @@ catalog:
   semver: 7.7.4
   winston: 3.19.0
   zod: 4.3.6
+  undici: 8.0.2
 
 minimumReleaseAge: 4320
