+++ sdk/core/package.json
@@ -47,7 +47,7 @@
     "@temporalio/workflow": "catalog:",
     "redis": "5.11.0",
     "stacktrace-parser": "0.1.11",
-    "undici": "8.0.2",
+    "undici": "catalog:",
     "winston": "catalog:",
     "zod": "catalog:"
   },
