+++ sdk/fetch/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@outputai/http2",
+  "name": "@outputai/fetch",
   "version": "0.1.11",
   "description": "Undici's fetch wrapper to trace http calls",
   "type": "module",
@@ -18,9 +18,7 @@
     "build": "mkdir -p dist && cp src/index.js src/utils.js src/config.js dist/"
   },
   "dependencies": {
-    "@outputai/core": "workspace:"
-  },
-  "devDependencies": {
+    "@outputai/core": "workspace:",
     "undici": "catalog:"
   },
   "license": "Apache-2.0",
