+++ sdk/fetch/src/utils.js
