+++ test_workflows/package.json
@@ -16,7 +16,7 @@
     "@outputai/credentials": "workspace:^",
     "@outputai/evals": "workspace:^",
     "@outputai/http": "workspace:^",
-    "@outputai/http2": "workspace:^",
+    "@outputai/fetch": "workspace:^",
     "@outputai/llm": "workspace:^",
     "@temporalio/workflow": "catalog:",
     "nodemon": "3.1.14"
